Added breedNextGeneration, start methods;
breedNextGeneration selects a subset of the current generation,
applies genetic operators to them (1 point crossover and mutation
at random positions in the current implementation) and creates
the next generation from these individuals.

start is the main loop of the GA; it calls breedNextGeneration until
the hasTerminated functor returns true, and then returns the individual
with the highest fitness from the last generation.